exVision/requirements.txt
matplotlib
numpy
PyQt5
pyqtdarktheme
qdarkstyle
scikit-image
scipy
Pillow
opencv-python-headless